requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

annotated-types==0.7.0
anyio==4.6.2.post1
azure-core==1.31.0
azure-functions==1.21.3
azure-storage-blob==12.23.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6
cryptography==43.0.3
distro==1.9.0
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
isodate==0.7.2
jiter==0.7.0
numpy==2.0.2
openai==1.54.3
pandas==2.2.3
pillow==11.0.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pyodbc==5.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
six==1.16.0
sniffio==1.3.1
tqdm==4.67.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
WooCommerce==3.0.0
lxml==5.3.0